Move model states to separate header files to avoid circular dependency
