+++ libsimulator/src/OperationalModels/CollisionFreeSpeedModel/CMakeLists.txt
@@ -2,5 +2,6 @@
 target_sources(simulator PRIVATE
     CollisionFreeSpeedModel.cpp
     CollisionFreeSpeedModel.hpp
+    CollisionFreeSpeedModelState.hpp
 )
 target_include_directories(simulator PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
