+++ libsimulator/src/OperationalModels/CollisionFreeSpeedModel/CollisionFreeSpeedModel.hpp
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 #pragma once
 
+#include "CollisionFreeSpeedModelState.hpp"
 #include "OperationalModel.hpp"
 #include "OperationalModelType.hpp"
 #include "Point.hpp"
@@ -13,13 +14,7 @@ struct WallView;
 class CollisionFreeSpeedModel : public OperationalModel
 {
 public:
-    /// Per-agent state of the collision free speed model.
-    struct State {
-        Point orientation{0.0, 0.0};
-        double timeGap{1};
-        double v0{1.2};
-        double radius{0.2};
-    };
+    using State = CollisionFreeSpeedModelState;
 
 private:
     double _cutOffRadius{3};
@@ -48,22 +43,4 @@ private:
     GetSpacing(const State& currState, const NeighborView& neighbor, const Point& direction) const;
     Point NeighborRepulsion(const State& currState, const NeighborView& neighbor) const;
     Point BoundaryRepulsion(const State& currState, const WallView& boundary) const;
-};
-
-template <>
-struct fmt::formatter<CollisionFreeSpeedModel::State> {
-
-    constexpr auto parse(format_parse_context& ctx) { return ctx.begin(); }
-
-    template <typename FormatContext>
-    auto format(const CollisionFreeSpeedModel::State& m, FormatContext& ctx) const
-    {
-        return fmt::format_to(
-            ctx.out(),
-            "CollisionFreeSpeedModel[orientation={}, timeGap={}, v0={}, radius={}])",
-            m.orientation,
-            m.timeGap,
-            m.v0,
-            m.radius);
-    }
 };
