+++ libsimulator/src/OperationalModels/CustomModel/CMakeLists.txt
@@ -1,4 +1,6 @@
 target_sources(simulator PRIVATE
     CustomModel.hpp
+    CustomModelState.hpp
     FormatAny.hpp
 )
+target_include_directories(simulator PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
