+++ libsimulator/src/OperationalModels/SocialForceModel/SocialForceModel.hpp
@@ -4,6 +4,7 @@
 #include "OperationalModel.hpp"
 #include "OperationalModelType.hpp"
 #include "Point.hpp"
+#include "SocialForceModelState.hpp"
 
 #include <fmt/core.h>
 
@@ -13,17 +14,7 @@ struct WallView;
 class SocialForceModel : public OperationalModel
 {
 public:
-    /// Per-agent state of the social force model.
-    struct State {
-        Point velocity{}; // v
-        double mass{80.0}; // m
-        double desiredSpeed{0.8}; // v0
-        double reactionTime{0.5}; // tau
-        double agentScale{2000.0}; // A for other agents
-        double obstacleScale{2000.0}; // A for obstacles
-        double forceDistance{0.08}; // B
-        double radius{0.3}; // r
-    };
+    using State = SocialForceModelState;
 
 private:
     double _cutOffRadius{2.5};
@@ -90,26 +81,4 @@ private:
      * @return length of pushing force between the two points
      */
     static double PushingForceLength(double A, double B, double r, double distance);
-};
-
-template <>
-struct fmt::formatter<SocialForceModel::State> {
-
-    constexpr auto parse(format_parse_context& ctx) { return ctx.begin(); }
-
-    template <typename FormatContext>
-    auto format(const SocialForceModel::State& m, FormatContext& ctx) const
-    {
-        return fmt::format_to(
-            ctx.out(),
-            "SFM[velocity={}, m={}, v0={}, tau={}, A_ped={}, A_obst={}, B={}, r={}])",
-            m.velocity,
-            m.mass,
-            m.desiredSpeed,
-            m.reactionTime,
-            m.agentScale,
-            m.obstacleScale,
-            m.forceDistance,
-            m.radius);
-    }
 };
